fixed display issue
